Remove wrapper class from lambdify
