+++ sympy/printing/lambdarepr.py
@@ -16,19 +16,6 @@ class LambdaPrinter(PythonCodePrinter):
     """
     printmethod = "_lambdacode"
 
-    def _print_MatrixBase(self, expr):
-        return "%s(%s)" % (expr.__class__.__name__,
-                           self._print(expr.tolist()))
-
-    _print_SparseMatrix = \
-        _print_MutableSparseMatrix = \
-        _print_ImmutableSparseMatrix = \
-        _print_Matrix = \
-        _print_DenseMatrix = \
-        _print_MutableDenseMatrix = \
-        _print_ImmutableMatrix = \
-        _print_ImmutableDenseMatrix = \
-        _print_MatrixBase
 
     def _print_And(self, expr):
         result = ['(']
@@ -205,7 +192,6 @@ class NumExprPrinter(LambdaPrinter):
             else:
                 raise TypeError("numexpr does not support function '%s'" %
                                 func_name)
-        print(nstr)
         return "%s(%s)" % (nstr, self._print_seq(e.args))
 
     def blacklisted(self, expr):
